decimals

+++ src/parser/parser.cpp
@@ -56,7 +56,7 @@ llvm::Value *parser::evaluateValue(std::string name, std::string value, std::siz
     }
 
     if (std::isdigit(value.c_str()[0])) {
-        llvm::Value *variable = llvm::ConstantFP::get(llvm::Type::getDoubleTy(*Context), (float)std::stoi(value));
+        llvm::Value *variable = llvm::ConstantFP::get(llvm::Type::getDoubleTy(*Context), std::stod(value));
         return variable;
     }
 
@@ -88,15 +88,15 @@ llvm::Value *parser::evaluateValue(std::string name, std::string value, std::siz
         }else if (lexedCode[originalIndex] == "div") {
             return Builder->CreateFDiv(first, second, "DivisionTemp");
         }else if (lexedCode[originalIndex] == "cmp") {
-        // Create the floating-point comparison
-        llvm::Value *FCmp = Builder->CreateFCmpOEQ(first, second, "FCMP");
+            // Create the floating-point comparison
+            llvm::Value *FCmp = Builder->CreateFCmpOEQ(first, second, "FCMP");
 
-        // Convert the i1 result (boolean) to i32 (32-bit integer)
-        llvm::Value *IntResult = Builder->CreateZExt(FCmp, llvm::Type::getInt32Ty(*Context), "zext");
+            // Convert the i1 result (boolean) to i32 (32-bit integer)
+            llvm::Value *IntResult = Builder->CreateZExt(FCmp, llvm::Type::getInt32Ty(*Context), "zext");
 
-        // Convert the i32 result to a double (64-bit floating point)
-        llvm::Value *DoubleResult = Builder->CreateSIToFP(IntResult, llvm::Type::getDoubleTy(*Context), "boolToDouble");
-        return DoubleResult;
+            // Convert the i32 result to a double (64-bit floating point)
+            llvm::Value *DoubleResult = Builder->CreateSIToFP(IntResult, llvm::Type::getDoubleTy(*Context), "boolToDouble");
+            return DoubleResult;
         }else if (lexedCode[originalIndex] == "pow") {
             return Builder->CreateCall(function->getFunction("pow"), {first, second}); // TODO: this is undefined
         }
